+++ sdk/identity/azure-identity/azure/identity/_credentials/chained.py
@@ -3,10 +3,10 @@
 # Licensed under the MIT License.
 # ------------------------------------
 import logging
-from typing import Any, Optional, TYPE_CHECKING
+from typing import Any, Optional, TYPE_CHECKING, cast
 from azure.core.exceptions import ClientAuthenticationError
 
-from azure.core.credentials import AccessToken
+from azure.core.credentials import AccessToken, AccessTokenInfo, TokenRequestOptions, SupportsTokenInfo
 from .. import CredentialUnavailableError
 from .._internal import within_credential_chain
 
@@ -74,6 +74,9 @@ class ChainedTokenCredential:
     ) -> AccessToken:
         """Request a token from each chained credential, in order, returning the first token received.
 
+        If no credential provides a token, raises :class:`azure.core.exceptions.ClientAuthenticationError`
+        with an error message from each credential.
+
         This method is called automatically by Azure SDK clients.
 
         :param str scopes: desired scopes for the access token. This method requires at least one scope.
@@ -104,6 +107,67 @@ class ChainedTokenCredential:
                 history.append((credential, str(ex)))
                 _LOGGER.debug(
                     '%s.get_token failed: %s raised unexpected error "%s"',
+                    self.__class__.__name__,
+                    credential.__class__.__name__,
+                    ex,
+                    exc_info=True,
+                )
+                break
+
+        within_credential_chain.set(False)
+        attempts = _get_error_message(history)
+        message = (
+            self.__class__.__name__
+            + " failed to retrieve a token from the included credentials."
+            + attempts
+            + "\nTo mitigate this issue, please refer to the troubleshooting guidelines here at "
+            "https://aka.ms/azsdk/python/identity/defaultazurecredential/troubleshoot."
+        )
+        _LOGGER.warning(message)
+        raise ClientAuthenticationError(message=message)
+
+    def get_token_info(self, *scopes: str, options: Optional[TokenRequestOptions] = None) -> AccessTokenInfo:
+        """Request a token from each chained credential, in order, returning the first token received.
+
+        If no credential provides a token, raises :class:`azure.core.exceptions.ClientAuthenticationError`
+        with an error message from each credential.
+
+        This is an alternative to `get_token` to enable certain scenarios that require additional properties
+        on the token. This method is called automatically by Azure SDK clients.
+
+        :param str scopes: desired scopes for the access token. This method requires at least one scope.
+            For more information about scopes, see https://learn.microsoft.com/entra/identity-platform/scopes-oidc.
+        :keyword options: A dictionary of options for the token request. Unknown options will be ignored. Optional.
+        :paramtype options: ~azure.core.credentials.TokenRequestOptions
+
+        :rtype: AccessTokenInfo
+        :return: An AccessTokenInfo instance containing information about the token.
+
+        :raises ~azure.core.exceptions.ClientAuthenticationError: no credential in the chain provided a token.
+        """
+        within_credential_chain.set(True)
+        history = []
+        for credential in self.credentials:
+            try:
+                # A custom credential in the chain may not implement get_token_info
+                if hasattr(credential, "get_token_info"):
+                    token_info = cast(SupportsTokenInfo, credential).get_token_info(*scopes, options=options)
+                else:
+                    options = options or {}
+                    token = credential.get_token(*scopes, **options)
+                    token_info = AccessTokenInfo(token=token.token, expires_on=token.expires_on)
+                _LOGGER.info("%s acquired a token from %s", self.__class__.__name__, credential.__class__.__name__)
+                self._successful_credential = credential
+                within_credential_chain.set(False)
+                return token_info
+            except CredentialUnavailableError as ex:
+                # credential didn't attempt authentication because it lacks required data or state -> continue
+                history.append((credential, ex.message))
+            except Exception as ex:  # pylint: disable=broad-except
+                # credential failed to authenticate, or something unexpectedly raised -> break
+                history.append((credential, str(ex)))
+                _LOGGER.debug(
+                    '%s.get_token_info failed: %s raised unexpected error "%s"',
                     self.__class__.__name__,
                     credential.__class__.__name__,
                     ex,